Fix proposals not appearing
