+++ src/components/api/api.ts
@@ -1,15 +1,17 @@
 import { address, abi } from "./contract.json";
 import ipfsClient from "ipfs-http-client";
 import Web3 from "web3";
-import { ProposalInfo } from "components/Vote/VoteDetail";
-
-const ipfsURL = "abcdao.infura-ipfs.io";
+import type { Contract } from "web3-eth-contract";
+import { Proposal, ProposalInfo } from "./types";
+const IPFS_URL = "abcdao.infura-ipfs.io";
 
 const ipfs = ipfsClient.create({
-  host: ipfsURL,
+  host: IPFS_URL,
   port: 5001,
   protocol: "https",
 });
+
+let contract: Contract | undefined;
 
 export async function uploadProposal(text: string) {
   const added = await ipfs.add(text);
@@ -19,15 +21,20 @@ export async function uploadProposal(text: string) {
 export async function retrieveProposal(
   proposalHash: string
 ): Promise<Proposal> {
-  return fetch("https://" + ipfsURL + "/" + proposalHash).then((x) => x.json());
+  return fetch(`https://${IPFS_URL}/ipfs/${proposalHash}`).then((x) =>
+    x.json()
+  );
 }
 
 export function getContract(web3: Web3) {
-  return new web3.eth.Contract(abi as any, address);
+  if (contract === undefined) {
+    contract = new web3.eth.Contract(abi as any, address);
+  }
+  return contract;
 }
 
 export async function getProposalHashes(web3: Web3): Promise<string[]> {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   return contract.methods.getProposalHashes().call();
 }
 
@@ -35,12 +42,12 @@ export async function getProposalInfo(
   web3: Web3,
   ipfsHash: string
 ): Promise<ProposalInfo> {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   return contract.methods.getProposalInfo(ipfsHash).call();
 }
 
 export async function getAllProposals(web3: Web3) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   return contract.methods.getAllProposals().call();
 }
 
@@ -49,7 +56,7 @@ export async function getVotesForOption(
   ipfsHash: string,
   optionIndex: number
 ) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   return contract.methods.getVotesForOption(ipfsHash, optionIndex).call();
 }
 
@@ -58,7 +65,7 @@ export async function initialiseUser(
   account: string,
   userAccount: string
 ) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   try {
     return await contract.methods
       .initialiseUser(userAccount)
@@ -68,29 +75,13 @@ export async function initialiseUser(
   }
 }
 
-export interface Proposal {
-  ipfs: string;
-  title: string;
-  content: string;
-  numOfOptions: number;
-  min_stake: number;
-  userId: string;
-  create_date: number;
-  type: string;
-  options: { id: number; label: string }[];
-  end_date: number;
-  isLossVoting: boolean;
-  isActive: boolean;
-  isAllocationProposal: boolean;
-}
-
 //can change anything except .create proposal
 export async function createProposal(
   web3: Web3,
   account: string | null,
   values: Proposal
 ) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   const ipfsHash = await uploadProposal(JSON.stringify(values));
   try {
     return await contract.methods
@@ -115,7 +106,7 @@ export async function setProposalStatus(
   ipfsHash: string,
   isActive: boolean
 ) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   try {
     return await contract.methods
       .setProposalStatus(ipfsHash, isActive)
@@ -132,7 +123,7 @@ export async function sendVote(
   optionIndex: number,
   stakeValue: number
 ) {
-  const contract = await getContract(web3);
+  const contract = getContract(web3);
   const amt = String(stakeValue * 1000000000000000000);
   try {
     return await contract.methods
